[FIX] translate: only delete migrated terms
After upgrading Odoo from earlier versions to versions >= Odoo 15.5, you
will notice that some fields that were translated before, are now in
English. This is likely because those translations were not marked as
`state='translated'` in the DB. However, the missing translations were
removed from the `_ir_translation` table during the upgrade.

With this patch, we still keep the conservative behavior of only
auto-translating fields marked as `state='translated'` in the DB (which
is still probably wrong, because user translations should remain the
same; if Odoo was displaying translated text before the upgrade and it
is not after it, that's probably a bug; but that's another story...),
but we no longer delete the other fields. This way, you can still mark
them as translated after the upgrade and use the new tooling provided in
https://github.com/odoo/upgrade-util/commit/fd578e31973596bf7f283ea5c6ebbdfd3314f8bc
to recover those translations after the upgrade.

@moduon MT-11570

closes odoo/odoo#228475

X-original-commit: e5e203d92a932b4168a297e1bbce3b5e7f4d5a47
Signed-off-by: Chong Wang (cwg) <cwg@odoo.com>
Co-authored-by: Alvaro Fuentes <4387571+aj-fuentes@users.noreply.github.com>